requirements.txt.txt
aiogram==3.0.0
python-dotenv
flask